bug(TCOMP-2309) Split file
